sx127x_netdev: TX power is int8_t (fixes HardFault on STM32L0)

+++ stack/framework/hal/chips/netdev_driver/sx127x_netdev.c
@@ -665,8 +665,8 @@ static int _set(netdev_t *netdev, netopt_t opt, const void *val, size_t len)
             return sizeof(uint32_t);
 
         case NETOPT_TX_POWER:
-            assert(len <= sizeof(int16_t));
-            int16_t power = *((const int16_t *)val);
+            assert(len <= sizeof(int8_t));
+            int8_t power = *((const int8_t *)val);
             if ((power < INT8_MIN) || (power > INT8_MAX)) {
                 res = -EINVAL;
                 break;
